🗃Refactor: TypeScript
