+++ packages/dhrod0325/src/core/types.ts
@@ -1,0 +1,1 @@
+export declare type Value<ValueType> = { [key: string | symbol]: ValueType };
